Init with slicing and MIG experiments
The MIG code checks MIG overhead for reconfiguring single-slice
instances. It is consistent with measurements from last year, but
inconsistent with those from a paper.

The slicing code attempts to measure the time quantum as well as the
time of a context switch. The latter is exhibiting weird behavior. I may
need to size the kernels to use all SMs to force full switches? These
are long-running thread blocks, so I have confirmed that switches happen
across instructions.